+++ packages/protocol-ethereum-sdk/src/order/sell.ts
@@ -1,30 +1,28 @@
 import type {
-	Address,
 	Erc20AssetType,
 	EthAssetType,
 	Order,
 	OrderForm,
-	Part,
+	RaribleV2OrderForm,
 } from "@rarible/protocol-api-client"
-import { randomWord, toBigNumber, Word } from "@rarible/types"
 import { BigNumberValue, toBn } from "@rarible/utils/build/bn"
 import { Action } from "@rarible/action"
-import type { AssetTypeRequest, AssetTypeResponse } from "./check-asset-type"
-import { UpsertOrder } from "./upsert-order"
+import { toBigNumber } from "@rarible/types"
+import type { HasOrder, HasPrice, OrderRequest, UpsertOrder } from "./upsert-order"
+import { AssetTypeRequest, AssetTypeResponse } from "./check-asset-type"
+import { SimpleOrder } from "./types"
+import { isCurrency } from "./is-currency"
 
 export type SellRequest = {
-	maker: Address
 	makeAssetType: AssetTypeRequest
 	amount: number
 	takeAssetType: EthAssetType | Erc20AssetType
-	price: BigNumberValue
-	payouts: Array<Part>
-	originFees: Array<Part>
-	salt?: Word
-}
-
-export type SellOrderAction = Action<SellOrderStageId, SellRequest, Order>
+} & HasPrice & OrderRequest
 export type SellOrderStageId = "approve" | "sign"
+export type SellOrderAction = Action<SellOrderStageId, SellRequest, Order>
+export type SellUpdateRequest = HasOrder & HasPrice
+
+export type SellOrderUpdateAction = Action<SellOrderStageId, SellUpdateRequest, Order>
 
 export class OrderSell {
 	constructor(
@@ -32,40 +30,64 @@ export class OrderSell {
 		private readonly checkAssetType: (asset: AssetTypeRequest) => Promise<AssetTypeResponse>,
 	) {}
 
-	sell: SellOrderAction = Action
+	readonly sell: SellOrderAction = Action
 		.create({
 			id: "approve" as const,
 			run: async (request: SellRequest) => {
-				const orderForm = await this.prepareOrderForm(request)
-				const checkedOrder = await this.upserter.checkLazyOrder(orderForm)
-				await this.upserter.approve(checkedOrder, false)
-				return checkedOrder
+				const form = await this.getSellForm(request)
+				const checked = await this.upserter.checkLazyOrder(form)
+				await this.upserter.approve(checked, false)
+				return checked
 			},
 		})
 		.thenStep({
 			id: "sign" as const,
-			run: async (checkedOrder: OrderForm) => this.upserter.upsertRequest(checkedOrder),
+			run: (form: OrderForm) => this.upserter.upsertRequest(form),
 		})
 
-	private async prepareOrderForm(request: SellRequest): Promise<OrderForm> {
-		const salt = request.salt ?? randomWord()
+	private async getSellForm(request: SellRequest): Promise<RaribleV2OrderForm> {
+		const price = await this.upserter.getPrice(request, request.takeAssetType)
+		const form = this.upserter.prepareOrderForm(request)
 		return {
-			maker: request.maker,
+			...form,
 			make: {
 				assetType: await this.checkAssetType(request.makeAssetType),
 				value: toBigNumber(request.amount.toString()),
 			},
 			take: {
 				assetType: request.takeAssetType,
-				value: toBigNumber(toBn(request.price).multipliedBy(request.amount).toString()),
+				value: toBigNumber(toBn(price).multipliedBy(request.amount).toString()),
 			},
-			type: "RARIBLE_V2",
-			data: {
-				dataType: "RARIBLE_V2_DATA_V1",
-				payouts: request.payouts,
-				originFees: request.originFees,
-			},
-			salt: toBigNumber(toBn(salt, 16).toString(10)) as any,
 		}
+	}
+
+	readonly update: SellOrderUpdateAction = Action
+		.create({
+			id: "approve" as const,
+			run: async (request: SellUpdateRequest) => {
+				const order = await this.upserter.getOrder(request)
+				if (!isCurrency(order.take.assetType)) {
+					throw new Error(`Not a sell order: ${JSON.stringify(order)}`)
+				}
+				const price = await this.upserter.getPrice(request, order.take.assetType)
+				const form = await this.prepareOrderUpdateForm(order, price)
+				const checked = await this.upserter.checkLazyOrder(form)
+				await this.upserter.approve(checked, false)
+				return checked
+			},
+		})
+		.thenStep({
+			id: "sign" as const,
+			run: (form: OrderForm) => this.upserter.upsertRequest(form),
+		})
+
+	async prepareOrderUpdateForm(order: SimpleOrder, price: BigNumberValue): Promise<OrderForm> {
+		if (order.type === "RARIBLE_V1" || order.type === "RARIBLE_V2") {
+			return this.upserter.getOrderFormFromOrder(order, order.make, {
+				assetType: order.take.assetType,
+				value: toBigNumber(toBn(price).multipliedBy(order.make.value).toString()),
+			})
+		}
+		throw new Error(`Unsupported order type: ${order.type}`)
 	}
 }